refactored the assignments list view page
