+++ frontend/src/components/assignments/AssignmentCard.tsx
@@ -1,32 +1,49 @@
-import { Card, Space, Tag, Typography } from 'antd';
+import { Card, Space, Typography } from 'antd';
 import { CalendarOutlined, ClockCircleOutlined, FileTextOutlined } from '@ant-design/icons';
 
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
 import type { Assignment } from '@/types/modules/assignments';
+import { useModule } from '@/context/ModuleContext';
+import { useNavigate } from 'react-router-dom';
+import { getRandomAssignmentStatus } from '@/constants/mock/assignment';
+import AssignmentStatusTag from './AssignmentStatusTag';
+import AssignmentTypeTag from './AssignmentTypeTag';
 dayjs.extend(relativeTime);
 
 const { Text, Paragraph } = Typography;
 
 interface Props {
   assignment: Assignment;
+  actions?: React.ReactNode[];
 }
 
-const AssignmentCard = ({ assignment }: Props) => {
+const AssignmentCard = ({ assignment, actions }: Props) => {
+  const navigate = useNavigate();
+  const module = useModule();
+
+  const handleClick = () => {
+    navigate(`/modules/${module.id}/assignments/${assignment.id}`);
+  };
+
   return (
     <Card
-      className="dark:bg-neutral-900 dark:border-none hover:shadow-lg transition-shadow duration-200"
+      className="dark:bg-neutral-900 dark:border-none hover:shadow-lg cursor-pointer transition-shadow duration-200"
+      onClick={handleClick}
       title={
         <div className="flex justify-between items-start">
           <div className="flex items-center gap-2">
             <FileTextOutlined className="text-lg text-blue-500" />
             <span className="font-medium text-base">{assignment.name}</span>
           </div>
-          <Tag color="blue" className="text-xs uppercase">
-            {assignment.assignment_type}
-          </Tag>
+
+          <div className="flex items-center gap-1">
+            <AssignmentTypeTag type={assignment.assignment_type} />
+            <AssignmentStatusTag status={getRandomAssignmentStatus()} />
+          </div>
         </div>
       }
+      actions={actions}
     >
       <Paragraph className="text-sm text-gray-700 dark:text-gray-300 mb-4">
         {assignment.description || 'No description provided.'}
