+++ frontend/src/constants/mock/assignment.ts
@@ -1,0 +1,6 @@
+import { ASSIGNMENT_STATUSES, type AssignmentStatus } from "@/types/modules/assignments";
+
+export function getRandomAssignmentStatus(): AssignmentStatus {
+  const i = Math.floor(Math.random() * ASSIGNMENT_STATUSES.length);
+  return ASSIGNMENT_STATUSES[i];
+}
